+++ src/index.css
@@ -8,9 +8,6 @@
     font-family: "Inter";
     font-size: 24px;
   }
-  div {
-    border: 1px solid black;
-  }
 }
 
 @layer components {
