--- src/pages/MatchList/components/ui/TdComp.tsx
@@ -1,7 +1,0 @@
-interface TdProps {
-  children: React.ReactNode;
-}
-
-const TdComp = ({ children }: TdProps) => <td className="border px-4 py-2">{children}</td>;
-
-export default TdComp;
